+++ src/commonMain/kotlin/csense/kotlin/extensions/ranges/IntRange.kt
@@ -17,9 +17,3 @@ public inline val IntRange.length: Int
  */
 public inline val IntRange.largest: Int
     get() = maxOf(last, start)
-
-/**
- * The maximum value in the range (exclusive)
- */
-public inline val IntRange.endExclusive: Int
-    get() = last + 1 //since  +1 is the next size thus above end inclusive is +1