+++ src/commonTest/kotlin/csense/kotlin/extensions/primitives/FloatTest.kt
@@ -114,4 +114,14 @@ class FloatTest {
         100.99f.decimalPart().assert(value = 0.99f, delta = 0.0001f)
         100.0005f.decimalPart().assert(value = 0.0005f, delta = 0.0001f)
     }
+
+    @Test
+    fun isNotNaN() {
+        Float.NaN.isNotNaN().assertFalse()
+        Float.MAX_VALUE.isNotNaN().assertTrue()
+        Float.MIN_VALUE.isNotNaN().assertTrue()
+        Float.NEGATIVE_INFINITY.isNotNaN().assertTrue()
+        Float.POSITIVE_INFINITY.isNotNaN().assertTrue()
+        0F.isNotNaN().assertTrue()
+    }
 }