+++ src/commonTest/kotlin/csense/kotlin/extensions/ranges/IntRangeTest.kt
@@ -34,17 +34,5 @@ class IntRangeTest {
 
     }
 
-    @Suppress("EmptyRange")
-    @Test
-    fun intRangeEndExclusive() {
-        val range1 = 0 until 20
-        range1.last.assert(19)
-        range1.endExclusive.assert(20)
 
-        val weirdRange = 40 until 41
-        weirdRange.endExclusive.assert(41)
-
-        val backwards = 20 until 10
-        backwards.endExclusive.assert(10)
-    }
 }