+++ src/commonTest/kotlin/csense/kotlin/patterns/ExpectedTest.kt
@@ -441,21 +441,20 @@ class ExpectedTest {
 
     }
 
-    @Test
-    fun expectedAction() {
+    class ExpectedAction {
 
         @Test
         fun actionThatReturnsFailed() {
-            val failed = expected<String, String>(action = {
+            val failed: Expected<String, String> = expected<String, String>(action = {
                 "failed".asFailed()
             })
             failed.assertIs<ExpectedFailed<String>>()
-            failed.error.assert("success")
+            failed.error.assert("failed")
         }
 
         @Test
         fun actionThatReturnsSuccess() {
-            val success = expected<String, String>(action = {
+            val success: Expected<String, String> = expected<String, String>(action = {
                 "success".asSuccess()
             })
             success.assertIs<ExpectedSuccess<String>>()
@@ -598,7 +597,6 @@ class ExpectedTest {
     }
 }
 
-class ErrorTypeException : Throwable()
 
 class ExpectedContextTest {
     @Test
@@ -623,3 +621,5 @@ class ExpectedContextTest {
         }.error.assert("hello")
     }
 }
+
+class ErrorTypeException : Throwable()