+++ changelog.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 0.0.60
+- targets java 11
+- tier 1 & 2 (mostly) + mingw native targets added
+
+- added
+    - RestartableJob
+    - floorMod
+    - Collection.onEmpty, Collection.onEmptyLazy
+    - 
+    - nullable deprecated methods for avoid "misuse" of nullable functions
+      (for all functions with involving nullable receivers and null as an action")
+
+
 ## 0.0.59
 
 - added